+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Improve UX of changing/updating group & contact names, add buttons in those dialogs to the dialog footer
 - Overhaul QR Code dialog and condense scan/show qr code into one dialog
 
+### Fixed
+
+- Fix sluggish ui while typing in email address in login form
+
+
 ## [1.12.0] - 2020-07-31
 
 ### Changed